requirements/project.txt
appnope==0.1.0
boto==2.38.0
coverage==3.7.1
csvkit==0.9.1
dbf==0.94.3
decorator==4.0.2
Django==1.8.4
django-appconf==1.0.1
django-bakery==0.7.5
django-compressor==1.5
django-debug-toolbar==1.3.2
django-extensions==1.5.6
django-postgres-copy==0.0.6
ecdsa==0.13
Fabric==1.10.2
gnureadline==6.3.3
gunicorn==19.3.0
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
jdcal==1.0
mimeparse==0.1.3
openpyxl==2.2.0b1
paramiko==1.15.2
path.py==8.1.1
pep8==1.6.2
pexpect==3.3
pickleshare==0.5
psycopg2==2.6.1
pycrypto==2.6.1
pyflakes==0.9.2
pystache==0.5.4
python-dateutil==2.2
python-memcached==1.57
simplegeneric==0.8.1
six==1.9.0
SQLAlchemy==1.0.8
sqlparse==0.1.16
static==1.1.1
traitlets==4.0.0
wheel==0.24.0
xlrd==0.9.4
